move addplayers dialog to runcampaign
