+++ src/components/campaign/Players.vue
@@ -30,7 +30,7 @@
 			</div>
 			<div class="d-flex justify-content-end">
 				<template v-if="viewerIsUser && page !== 'user'">
-					<button class="btn btn-sm bg-neutral-5 mr-1" @click="add_players_dialog = true">
+					<button class="btn btn-sm bg-neutral-5 mr-1" @click="open_player_dialog">
 						<i aria-hidden="true" class="fas fa-user"></i>
 						<q-tooltip anchor="top middle" self="center middle">Manage Players</q-tooltip>
 					</button>
@@ -370,11 +370,6 @@
 		</div>
 
 		<q-resize-observer @resize="onResize" />
-
-		<!-- Edit campaign dialog -->
-		<q-dialog v-if="!overencumbered" v-model="add_players_dialog">
-			<AddPlayers :campaign="search_campaign" />
-		</q-dialog>
 	</tag>
 </template>
 
@@ -382,7 +377,6 @@
 import { mapGetters, mapActions } from "vuex";
 import { experience } from "src/mixins/experience.js";
 import { currencyMixin } from "src/mixins/currency.js";
-import AddPlayers from "src/components/campaign/AddPlayers";
 
 export default {
 	name: "Players",
@@ -405,7 +399,6 @@ export default {
 			default: false,
 		},
 	},
-	components: { AddPlayers },
 	mixins: [experience, currencyMixin],
 	data() {
 		return {
@@ -417,7 +410,6 @@ export default {
 			loading: false,
 			isXpAdvancement: false,
 			add_players_dialog: false,
-			search_campaign: {},
 		};
 	},
 	computed: {
@@ -526,6 +518,9 @@ export default {
 		maxHp(maxHp, maxHpMod) {
 			return parseInt(maxHpMod ? maxHp + maxHpMod : maxHp);
 		},
+		open_player_dialog() {
+			this.$emit("add-player", true);
+		},
 		reset() {
 			for (const [id, player] of Object.entries(this.players)) {
 				if (!player.dead) {
@@ -567,25 +562,6 @@ export default {
 				}
 			}
 		},
-	},
-	async mounted() {
-		const properties = [
-			"name",
-			"background",
-			"hk_background",
-			"player_count",
-			"advancement",
-			"timestamp",
-			"private",
-		];
-
-		this.search_campaign = { key: this.campaignId };
-		for (const prop of properties) {
-			if (this.campaign.hasOwnProperty(prop)) {
-				this.search_campaign[prop] = this.campaign[prop];
-			}
-		}
-		console.log(this.search_campaign, this.campaign);
 	},
 };
 </script>
